RasPiManager/requirements.txt
# requirements.txt

# Python 3.11.X'
RPi.GPIO
rich
open_gopro
smbus2